# crates/manta-cli/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manta API",
    "description": "REST API for managing CSM/OpenCHAMI HPC clusters via manta.",
    "contact": {
      "name": "Manuel Sopena Ballesteros",
      "email": "msopena@cscs.ch"
    },
    "license": {
      "name": ""
    },
    "version": "2.0.0-beta.64"
  },
  "servers": [
    {
      "url": "/api/v2",
      "description": "manta API v2"
    }
  ],
  "paths": {
...
    "/templates": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "GET /templates — list BOS session templates with optional filters.",
        "operationId": "get_templates",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Exact template name.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "hsm_group",
            "in": "query",
            "description": "HSM group whose associated templates should be returned.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Cap on the number of templates returned. BOS templates have no\ntimestamp, so this caps the count of the name-sorted list; it does\nnot select by recency.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of session templates",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/templates/{name}/sessions": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "`POST /api/v2/templates/{name}/sessions` — create a BOS session from a session template.",
        "operationId": "post_template_session",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Template name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostTemplateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Dry run preview",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "201": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "BosOperation": {
        "type": "string",
        "description": "BOS session operation to run against the template's node list.",
        "enum": [
          "boot",
          "reboot",
          "shutdown"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard JSON error body returned by all failed endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable explanation of the failure. Never includes\nstack traces, credentials, or internal type names."
          }
        }
      },
...
      "PostTemplateSessionRequest": {
        "type": "object",
        "description": "Request body for `POST /api/v2/templates/{name}/sessions`.\n\nCreates a BOS session from an existing session template (the\n`{name}` path segment).\n\n# Wire shape\n\n```json\n{\n  \"operation\": \"reboot\",\n  \"limit\": \"x3000c0s1b0n[0-3]\",\n  \"session_name\": null,\n  \"include_disabled\": false,\n  \"dry_run\": false\n}\n```",
        "required": [
          "operation",
          "limit"
        ],
        "properties": {
          "dry_run": {
            "type": "boolean",
            "description": "When true, validate the session parameters without creating a\nBOS session."
          },
          "include_disabled": {
            "type": "boolean",
            "description": "When true, include nodes marked as disabled."
          },
          "limit": {
            "type": "string",
            "description": "Ansible limit expression restricting which template nodes are\ntargeted."
          },
          "operation": {
            "$ref": "#/components/schemas/BosOperation",
            "description": "BOS operation to run (boot, reboot, or shutdown)."
          },
          "session_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional explicit name for the BOS session; auto-generated when\nabsent."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}