# crates/manta-shared/src/types/api/template.rs
//! HTTP request/response bodies and CLI-built parameter structs for
//! the BOS session template endpoints (`/api/v2/templates`,
//! `/api/v2/templates/{name}/sessions`).

use serde::{Deserialize, Serialize};
use utoipa::ToSchema;

/// BOS session operation to run against the template's node list.
#[derive(Debug, Clone, Copy, Serialize, Deserialize, ToSchema)]
#[serde(rename_all = "lowercase")]
pub enum BosOperation {
  /// Boot nodes that are currently off.
  Boot,
  /// Reboot (power-cycle) nodes.
  Reboot,
  /// Shut down nodes.
  Shutdown,
}

impl BosOperation {
  /// Wire-level string form expected by the backend (`"boot"` /
  /// `"reboot"` / `"shutdown"`).
  pub fn as_str(&self) -> &'static str {
    match self {
      Self::Boot => "boot",
      Self::Reboot => "reboot",
      Self::Shutdown => "shutdown",
    }
  }
}

/// Request body for `POST /api/v2/templates/{name}/sessions`.
///
/// Creates a BOS session from an existing session template (the
/// `{name}` path segment).
///
/// # Wire shape
///
/// ```json
/// {
///   "operation": "reboot",
///   "limit": "x3000c0s1b0n[0-3]",
///   "session_name": null,
///   "include_disabled": false,
///   "dry_run": false
/// }
/// ```
#[derive(Debug, Serialize, Deserialize, ToSchema)]
pub struct PostTemplateSessionRequest {
  /// BOS operation to run (boot, reboot, or shutdown).
  pub operation: BosOperation,
  /// Ansible limit expression restricting which template nodes are
  /// targeted.
  pub limit: String,
  /// Optional explicit name for the BOS session; auto-generated when
  /// absent.
  pub session_name: Option<String>,
  /// When true, include nodes marked as disabled.
  #[serde(default)]
  pub include_disabled: bool,
  /// When true, validate the session parameters without creating a
  /// BOS session.
  #[serde(default)]
  pub dry_run: bool,
}

/// Typed parameters for fetching BOS session templates.
pub struct GetTemplateParams {
  /// Exact template name.
  pub name: Option<String>,
  /// Group whose associated templates should be returned.
  pub group_name: Option<String>,
  /// Operator default from `cli.toml`'s `hsm_group`, used
  /// as a fallback for `group_name`.
  pub settings_group_name: Option<String>,
  /// Cap on the number of templates returned. BOS templates have no
  /// timestamp, so this caps the count of the name-sorted list; it does
  /// not select by recency.
  pub limit: Option<u8>,
}

impl GetTemplateParams {
  /// Returns the effective group name, preferring the explicit
  /// `--group` flag and falling back to the operator default from
  /// `cli.toml`.
  pub fn effective_group(&self) -> Option<&str> {
    self
      .group_name
      .as_deref()
      .or(self.settings_group_name.as_deref())
  }
}

/// Parameters for applying a BOS session template.
pub struct ApplyTemplateParams {
  /// Optional explicit name for the created BOS session;
  /// auto-generated when absent.
  pub bos_session_name: Option<String>,
  /// Name of an existing BOS session template to instantiate.
  pub bos_sessiontemplate_name: String,
  /// Operation to perform: `"boot"`, `"reboot"`, or `"shutdown"`.
  pub bos_session_operation: String,
  /// Ansible-style limit expression scoping which template nodes
  /// are targeted (xnames, NIDs, groups, or roles).
  pub limit: String,
  /// When true, include nodes marked as disabled in the hardware
  /// state manager.
  pub include_disabled: bool,
}